REF: Create geometry submodule with classes in separate file
The geometry file is very large. We are going to shorten it by
breaking it into smaller files that are more topical. This
makes navigation easier.
