+++ src/xdesign/geometry/__init__.py
@@ -1,0 +1,17 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""Define geometric objects for compuational geometry operations.
+
+.. moduleauthor:: Doga Gursoy <dgursoy@aps.anl.gov>
+.. moduleauthor:: Daniel J Ching <carterbox@users.noreply.github.com>
+"""
+
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
+from xdesign.geometry.algorithms import *
+from xdesign.geometry.area import *
+from xdesign.geometry.entity import *
+from xdesign.geometry.line import *
+from xdesign.geometry.point import *
