+++ src/xdesign/geometry/area.py
@@ -1,56 +1,3 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-# #########################################################################
-# Copyright (c) 2016, UChicago Argonne, LLC. All rights reserved.         #
-#                                                                         #
-# Copyright 2016. UChicago Argonne, LLC. This software was produced       #
-# under U.S. Government contract DE-AC02-06CH11357 for Argonne National   #
-# Laboratory (ANL), which is operated by UChicago Argonne, LLC for the    #
-# U.S. Department of Energy. The U.S. Government has rights to use,       #
-# reproduce, and distribute this software.  NEITHER THE GOVERNMENT NOR    #
-# UChicago Argonne, LLC MAKES ANY WARRANTY, EXPRESS OR IMPLIED, OR        #
-# ASSUMES ANY LIABILITY FOR THE USE OF THIS SOFTWARE.  If software is     #
-# modified to produce derivative works, such modified software should     #
-# be clearly marked, so as not to confuse it with the version available   #
-# from ANL.                                                               #
-#                                                                         #
-# Additionally, redistribution and use in source and binary forms, with   #
-# or without modification, are permitted provided that the following      #
-# conditions are met:                                                     #
-#                                                                         #
-#     * Redistributions of source code must retain the above copyright    #
-#       notice, this list of conditions and the following disclaimer.     #
-#                                                                         #
-#     * Redistributions in binary form must reproduce the above copyright #
-#       notice, this list of conditions and the following disclaimer in   #
-#       the documentation and/or other materials provided with the        #
-#       distribution.                                                     #
-#                                                                         #
-#     * Neither the name of UChicago Argonne, LLC, Argonne National       #
-#       Laboratory, ANL, the U.S. Government, nor the names of its        #
-#       contributors may be used to endorse or promote products derived   #
-#       from this software without specific prior written permission.     #
-#                                                                         #
-# THIS SOFTWARE IS PROVIDED BY UChicago Argonne, LLC AND CONTRIBUTORS     #
-# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT       #
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS       #
-# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL UChicago     #
-# Argonne, LLC OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,        #
-# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,    #
-# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;        #
-# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER        #
-# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT      #
-# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN       #
-# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE         #
-# POSSIBILITY OF SUCH DAMAGE.                                             #
-# #########################################################################
-"""Defines geometric objects to support :class:`.Phantom` definition and
-perform compuational geometry for :mod:`.acquisition`.
-
-.. moduleauthor:: Doga Gursoy <dgursoy@aps.anl.gov>
-.. moduleauthor:: Daniel J Ching <carterbox@users.noreply.github.com>
-"""
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
@@ -65,543 +12,24 @@ import copy
 from math import sqrt, asin
 from copy import deepcopy
 
+from xdesign.geometry.entity import *
+from xdesign.geometry.point import *
+from xdesign.geometry.line import *
+
 logger = logging.getLogger(__name__)
 
 __author__ = "Daniel Ching, Doga Gursoy"
 __copyright__ = "Copyright (c) 2016, UChicago Argonne, LLC."
 __docformat__ = 'restructuredtext en'
-__all__ = ['Entity',
-           'Point',
+__all__ = [
+           'Curve',
            'Circle',
-           'Line',
            'Polygon',
            'Triangle',
            'Rectangle',
            'Square',
-           'Mesh'
+           'Mesh',
            ]
-
-
-class Entity(object):
-    """Base class for all geometric entities. All geometric entities should
-    have these attributes and methods.
-
-    Example
-    -------
-    Examples can be given using either the ``Example`` or ``Examples``
-    sections. Sections support any reStructuredText formatting, including
-    literal blocks::
-
-        $ python example_numpy.py
-
-
-    Section breaks are created with two blank lines. Section breaks are also
-    implicitly created anytime a new section starts. Section bodies *may* be
-    indented:
-
-    Parameters
-    ----------
-    x : :class:`.ndarray`, :class:`.list`
-        ND coordinates of the point.
-
-    Notes
-    -----
-        This is an example of an indented section. It's like any other section,
-        but the body is indented to help it stand out from surrounding text.
-
-    If a section is indented, then a section break is created by
-    resuming unindented text.
-
-
-    .. note::
-        There are many other directives such as versionadded, versionchanged,
-        rubric, centered, ... See the sphinx documentation for more details.
-
-    """
-
-    def __init__(self):
-        self._dim = 0
-
-    def __repr__(self):
-        """A string representation for easier debugging.
-
-        .. note::
-            This method is inherited from :class:`.Entity` which means it is
-            not implemented and will throw an error.
-        """
-        raise NotImplementedError
-
-    @property
-    def dim(self):
-        """The dimensionality of the points which describe the entity.
-        """
-        return self._dim
-
-    def translate(self, vector):
-        """Translates the entity in the direction of a vector.
-
-        .. note::
-            This method is inherited from :class:`.Entity` which means it is
-            not implemented and will throw an error.
-        """
-        raise NotImplementedError
-
-    def rotate(self, theta, point=None, axis=None):
-        """Rotates the entity theta radians around an axis defined by a point
-        and a vector
-
-        .. note::
-            This method is inherited from :class:`.Entity` which means it is
-            not implemented and will throw an error.
-        """
-        raise NotImplementedError
-
-    def scale(self, vector):
-        """Scales the entity in each dimension according to vector. Scaling is
-        centered on the origin.
-
-        .. note::
-            This method is inherited from :class:`.Entity` which means it is
-            not implemented and will throw an error.
-        """
-        raise NotImplementedError
-
-    def contains(self, other):
-        """Return whether this Entity strictly contains the other entity.
-
-        Points on edges are contained by the Entity.
-
-        Returns a boolean for all :class:`Entitity`. Returns an array of
-        boolean for MxN size arrays where M is the number of points and N is
-        the dimensionality.
-
-        .. note::
-            This method is inherited from :class:`.Entity` which means it is
-            not implemented and will throw an error.
-        """
-        raise NotImplementedError
-
-    def collision(self, other):
-        """Returns True if this entity collides with another entity.
-
-        .. note::
-            This method is inherited from :class:`.Entity` which means it is
-            not implemented and will throw an error.
-        """
-        raise NotImplementedError
-
-    def distance(self, other):
-        """Returns the closest distance between entities.
-
-        .. note::
-            This method is inherited from :class:`.Entity` which means it is
-            not implemented and will throw an error.
-        """
-        raise NotImplementedError
-
-    def midpoint(self, other):
-        """Returns the midpoint between entities."""
-        return self.distance(other) / 2.
-
-
-class Point(Entity):
-    """A point in ND cartesian space.
-
-    Parameters
-    ----------
-    x : :class:`.ndarray`, :class:`.list`
-        ND coordinates of the point.
-
-    Raises
-    ------
-    TypeError
-        If x is not a list or ndarray.
-    """
-    def __init__(self, x):
-        if not isinstance(x, (list, np.ndarray)):
-            raise TypeError("x must be list, or array of coordinates.")
-
-        super(Point, self).__init__()
-        self._x = np.array(x, dtype=float, ndmin=1)
-        self._x = np.ravel(self._x)
-        self._dim = self._x.size
-
-    def __repr__(self):
-        return "Point([%s" % ', '.join([repr(n) for n in self._x]) + "])"
-
-    @property
-    def x(self):
-        """Dimension 0 of the point."""
-        return self._x[0]
-
-    @property
-    def y(self):
-        """Dimension 1 of the point."""
-        return self._x[1]
-
-    @property
-    def z(self):
-        """Dimension 2 of the point."""
-        return self._x[2]
-
-    @property
-    def norm(self):
-        """Calculates the euclidian (L2) norm of the vector to the point."""
-        # See http://stackoverflow.com/a/23576322 for a discussion of the
-        # quickest way to calculate the norm of a vector.
-        return sqrt(self._x.dot(self._x))
-
-    def translate(self, vector):
-        """Translates the point along the given vector."""
-        if not isinstance(vector, (list, np.ndarray)):
-            raise TypeError("vector must be arraylike.")
-
-        self._x += vector
-
-    def rotate(self, theta, point=None, axis=None):
-        """Rotates the point theta radians around the axis defined by the given
-        point and axis."""
-        if not isinstance(theta, Number):
-            raise TypeError("theta must be scalar.")
-        if point is None:
-            center = np.zeros(self.dim)
-        elif isinstance(point, Point):
-            center = point._x
-        else:
-            raise TypeError("center of rotation must be Point.")
-        if axis is not None:
-            raise NotImplementedError("Rotation about axis besides [0 0 1] are"
-                                      " not implemented.")
-
-        # shift rotation center to origin
-        self._x -= center
-
-        # do rotation
-        R = np.eye(self.dim)
-        R[0, 0] = np.cos(theta)
-        R[0, 1] = -np.sin(theta)
-        R[1, 0] = np.sin(theta)
-        R[1, 1] = np.cos(theta)
-
-        self._x = np.dot(R, self._x)
-        # shift rotation center back
-        self._x += center
-
-    def scale(self, vector):
-        """Scales the coordinates of the point in each dimension according to
-        the given vector. Scaling is centered on the origin."""
-        if not isinstance(vector, (List, np.ndarray)):
-            raise TypeError("vector must be arraylike.")
-
-        self._x *= vector
-
-    def contains(self, other):
-        """Return wether the other is within the bounds of the Point. Points
-        can only contain other Points."""
-        if isinstance(other, Point):
-            return self == point
-        if isinstance(other, np.ndarray):
-            return np.all(self._x == other, axis=1)
-        else:  # points can only contain points
-            return False
-
-    def collision(self, other):
-        """Returns True if this Point collides with another entity."""
-        if isinstance(other, Point):
-            return self == point
-        else:
-            raise NotImplementedError
-
-    def distance(self, other):
-        """Returns the closest distance between entities."""
-        if isinstance(other, LinearEntity):
-            return other.distance(self)
-        elif not isinstance(other, Point):
-            raise NotImplementedError("Point to point distance only.")
-        d = self._x - other._x
-        return sqrt(d.dot(d))
-
-    # OVERLOADS
-    def __eq__(self, point):
-        if not isinstance(point, Point):
-            raise TypeError("Points can only equal to other points.")
-        return np.array_equal(self._x, point._x)
-
-    def __add__(self, point):
-        """Addition."""
-        if not isinstance(point, Point):
-            raise TypeError("Points can only add to other points.")
-        return Point(self._x + point._x)
-
-    def __sub__(self, point):
-        """Subtraction."""
-        if not isinstance(point, Point):
-            raise TypeError("Points can only subtract from other points.")
-        return Point(self._x - point._x)
-
-    def __mul__(self, c):
-        """Scalar, vector multiplication."""
-        if not isinstance(c, Number):
-            raise TypeError("Points can only multiply scalars.")
-        return Point(self._x * c)
-
-    def __truediv__(self, c):
-        """Scalar, vector division."""
-        if not isinstance(c, Number):
-            raise TypeError("Points can only divide scalars.")
-        return Point(self._x / c)
-
-    def __hash__(self):
-        return hash(self._x[:])
-
-
-class LinearEntity(Entity):
-    """Base class for linear entities in 2D Cartesian space. e.g. :class:`.Line`,
-    :class:`.Segment`, and :class:`.Ray`.
-
-    The constructor takes two unique :class:`.Point`.
-
-    Attributes
-    ----------
-    p1 : Point
-    p2 : Point
-    """
-    def __init__(self, p1, p2):
-        if not isinstance(p1, Point) or not isinstance(p2, Point):
-            raise TypeError("p1 and p2 must be Points")
-        if p1 == p2:
-            raise ValueError('Requires two unique Points.')
-        if p1.dim != p2.dim:
-            raise ValueError('Two Points must have same dimensionality.')
-        self.p1 = p1
-        self.p2 = p2
-        self._dim = p1.dim
-
-    def __repr__(self):
-        return "{}({}, {})".format(type(self).__name__, repr(self.p1),
-                                   repr(self.p2))
-
-    @property
-    def vertical(self):
-        """True if line is vertical."""
-        return self.p1.x == self.p2.x
-
-    @property
-    def horizontal(self):
-        """True if line is horizontal."""
-        return self.p1.y == self.p2.y
-
-    @property
-    def slope(self):
-        """Returns the slope of the line."""
-        if self.vertical:
-            return np.inf
-        else:
-            return ((self.p2.y - self.p1.y) /
-                    (self.p2.x - self.p1.x))
-
-    @property
-    def points(self):
-        """Returns the two points used to define this linear entity as a
-        2-tuple."""
-        return (self.p1, self.p2)
-
-    # @property
-    # def bounds(self):
-    #     """Return a tuple (xmin, ymin, xmax, ymax) representing the
-    #     bounding rectangle for the geometric figure.
-    #     """
-    #     xs = [p.x for p in self.points]
-    #     ys = [p.y for p in self.points]
-    #     return (min(xs), min(ys), max(xs), max(ys))
-
-    @property
-    def length(self):
-        """Returns the length of the segment between p1 and p2."""
-        return self.p1.distance(self.p2)
-
-    @property
-    def tangent(self):
-        """Returns the unit tangent vector."""
-        dx = (self.p2._x - self.p1._x) / self.length
-        return Point(dx)
-
-    @property
-    def normal(self):
-        """Return the unit normal vector."""
-        dx = (self.p2._x - self.p1._x) / self.length
-        R = np.array([[0, 1],
-                      [-1,  0]])
-        n = np.dot(R, dx)
-        return Point(n)
-
-    @property
-    def numpy(self):
-        """Returns an array of coordinates where the first row is p1 and the
-        second row is p2."""
-        return np.stack((self.p1._x, self.p2._x), axis=0)
-
-    @property
-    def list(self):
-        """Returns an list of coordinates where p1 is the first D coordinates
-        and p2 is the next D coordinates."""
-        return np.concatenate((self.p1._x, self.p2._x), axis=0)
-
-    def translate(self, vector):
-        """Translates the :class:`.LinearEntity` by the given vector."""
-        self.p1.translate(vector)
-        self.p2.translate(vector)
-
-    def rotate(self, theta, point=None, axis=None):
-        """Rotates the :class:`.LinearEntity` by theta radians around an axis
-        defined by an axis and a point."""
-        self.p1.rotate(theta, point, axis)
-        self.p2.rotate(theta, point, axis)
-
-
-class Line(LinearEntity):
-    """Line in 2D cartesian space.
-
-    The constructor takes two unique :class:`.Point`.
-
-    Attributes
-    ----------
-    p1 : Point
-    p2 : Point
-    """
-
-    def __init__(self, p1, p2):
-        super(Line, self).__init__(p1, p2)
-
-    def __str__(self):
-        """Return line equation."""
-        if self.vertical:
-            return "x = %s" % self.p1.x
-        elif self.dim == 2:
-            return "y = %sx + %s" % (self.slope, self.yintercept)
-        else:
-            A, B = self.standard
-            return "%sx " % '+ '.join([str(n) for n in A]) + "= " + str(B)
-
-    def __eq__(self, line):
-        return (self.slope, self.yintercept) == (line.slope, line.yintercept)
-
-    def intercept(self, n):
-        """Calculates the intercept for the nth dimension."""
-        if n > self._dim:
-            return 0
-        else:
-            A, B = self.standard
-            if A[n] == 0:
-                return np.inf
-            else:
-                return B/A[n]
-
-    @property
-    def xintercept(self):
-        """Return the x-intercept."""
-        if self.horizontal:
-            return np.inf
-        else:
-            return self.p1.x - 1 / self.slope * self.p1.y
-
-    @property
-    def yintercept(self):
-        """Return the y-intercept."""
-        if self.vertical:
-            return np.inf
-        else:
-            return self.p1.y - self.slope * self.p1.x
-
-    @property
-    def standard(self):
-        """Returns coeffients for the first N-1 standard equation coefficients.
-        The Nth is returned separately."""
-        A = np.stack([self.p1._x, self.p2._x], axis=0)
-        return calc_standard(A)
-
-    def distance(self, other):
-        """Returns the closest distance between entities."""
-        # REF: http://geomalgorithms.com/a02-_lines.html
-        if not isinstance(other, Point):
-            raise NotImplementedError("Line to point distance only.")
-        d = np.cross(self.tangent._x, other._x - self.p1._x)
-        if self.dim > 2:
-            return sqrt(d.dot(d))
-        else:
-            return abs(d)
-
-
-class Ray(Line):
-    """Ray in 2-D cartesian space.
-
-    It is defined by two distinct points.
-
-    Attributes
-    ----------
-    p1 : Point (source)
-    p2 : Point (point direction)
-    """
-
-    def __init__(self, p1, p2):
-        super(Ray, self).__init__(p1, p2)
-
-    @property
-    def source(self):
-        """The point from which the ray emanates."""
-        return self.p1
-
-    @property
-    def direction(self):
-        """The direction in which the ray emanates."""
-        return self.p2 - self.p1
-
-    def distance(self, other):
-        # REF: http://geomalgorithms.com/a02-_lines.html
-        v = self.p2._x - self.p1._x
-        w = other._x - self.p1._x
-
-        c1 = np.dot(w, v)
-
-        if c1 <= 0:
-            return self.p1.distance(other)
-        else:
-            return super(Ray, self).distance(other)
-
-
-class Segment(Line):
-    """Segment in 2-D cartesian space.
-
-    It is defined by two distinct points.
-
-    Attributes
-    ----------
-    p1 : Point
-    p2 : Point
-    """
-
-    def __init__(self, p1, p2):
-        super(Segment, self).__init__(p1, p2)
-
-    @property
-    def midpoint(self):
-        """The midpoint of the line segment."""
-        return Point.midpoint(self.p1, self.p2)
-
-    def distance(self, other):
-        # REF: http://geomalgorithms.com/a02-_lines.html
-        v = self.p2._x - self.p1._x
-        w = other._x - self.p1._x
-
-        c1 = np.dot(w, v)
-        c2 = np.dot(v, v)
-
-        if c1 <= 0:
-            return self.p1.distance(other)
-        elif c2 <= c1:
-            return self.p2.distance(other)
-        else:
-            return super(Segment, self).distance(other)
 
 
 class Curve(Entity):
@@ -1326,136 +754,3 @@ class Mesh(Entity):
         for f in self.faces:
             patches.append(f.patch)
         return patches
-
-
-def calc_standard(A):
-    """Return the standard equation (c_{0}*x + ... = c_{1}) coefficents for the
-    hyper-plane defined by the row-wise N-dimensional points A.
-
-    Parameters
-    ----------
-    A : :py:class:`np.array` (..., N, N)
-        Each row is an N-dimensional point on the plane.
-
-    Returns
-    ----------
-    c0 : :py:class:`np.array` (..., N)
-        The first N coeffients for the hyper-plane
-    c1 : :py:class:`np.array` (..., 1)
-        The last coefficient for the hyper-plane
-    """
-    b = np.ones(A.shape[0:-1])
-    x1 = np.atleast_1d(b[..., 0])
-    try:
-        x = np.linalg.solve(A, b)
-    except np.linalg.LinAlgError as e:
-        if str(e) != 'Singular matrix':
-            raise
-        else:
-            a = A[...,1,1] - A[...,0,1]
-            b = A[...,0,0] - A[...,1,0]
-            x1 = np.atleast_1d(a * A[...,0,0] + b * A[...,0,1])
-            x = np.stack([a, b], axis=-1)
-    return x, x1
-
-
-def halfspacecirc(d, r):
-    """Area of intersection between circle and half-plane. Returns the smaller
-    fraction of a circle split by a line d units away from the center of the
-    circle.
-
-    Reference
-    ---------
-    Glassner, A. S. (Ed.). (2013). Graphics gems. Elsevier.
-
-    Parameters
-    ----------
-    d : scalar
-        The distance from the line to the center of the circle
-    r : scalar
-        The radius of the circle
-
-    Returns
-    -------
-    f : scalar
-        The proportion of the circle in the smaller half-space
-    """
-    assert r > 0, "The radius must positive"
-    assert d >= 0, "The distance must be positive or zero."
-
-    if d >= r:  # The line is too far away to overlap!
-        return 0
-
-    f = 0.5 - d*sqrt(r**2 - d**2)/(np.pi*r**2) - asin(d/r)/np.pi
-
-    # Returns the smaller fraction of the circle, so it can be at most 1/2.
-    if f < 0 or 0.5 < f:
-        if f < -1e-16:  # f will often be less than 0 due to rounding errors
-            warnings.warn("halfspacecirc was out of bounds, {}".format(f),
-                          RuntimeWarning)
-        f = 0
-
-    return f
-
-
-def two_lines_intersect(l0A, l0b, l1A, l1b):
-    A = np.stack([l0A, l1A], axis=0)
-    b = np.stack([l0b, l1b], axis=0)
-    x = np.linalg.solve(A, b)
-    return x
-
-
-def half_space(self, center):
-    """Returns the half space polytope respresentation of the Line."""
-    A, B = self.standard
-    # test for positive or negative side of line
-    if not halfspace_has_point(A, B, center):
-        A = -A
-        B = -B
-
-    return A, B
-
-
-def halfspace_has_point(A, B, point):
-    return np.dot(A, point._x) <= B
-
-
-def clip_SH(clipEdges, polygon):
-    """Sutherland-Hodgeman algorithm for clipping a polygon
-
-    Parameters
-    ----------
-    clipEdges [[A, b], ...]
-        half-spaces defined by coefficients
-
-    polygon
-
-    """
-    outputList = polygon.vertices
-    for clipEdge in clipEdges:
-        # previous iteration output is this iteration input
-        inputList = outputList
-        outputList = list()
-        if len(inputList) == 0:
-            break
-        S = inputList[-1]
-
-        for E in inputList:
-
-            if halfspace_has_point(clipEdge[0], clipEdge[1], E):
-
-                if not halfspace_has_point(clipEdge[0], clipEdge[1], S):
-                    A, b = calc_standard(np.stack([S._x, E._x], axis=0))
-                    new_vert = two_lines_intersect(A, b,
-                                                   clipEdge[0], clipEdge[1])
-                    outputList.append(Point(new_vert))
-
-                outputList.append(E)
-
-            elif halfspace_has_point(clipEdge[0], clipEdge[1], S):
-                A, b = calc_standard(np.stack([S._x, E._x], axis=0))
-                new_vert = two_lines_intersect(A, b, clipEdge[0], clipEdge[1])
-                outputList.append(Point(new_vert))
-
-            S = E
-    return outputList
