+++ tests/geometry/test_area.py
@@ -50,211 +50,13 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
 from xdesign.geometry import *
-from xdesign.acquisition import beamcirc, beampoly
-from xdesign.acquisition import *
-from numpy.testing import assert_allclose, assert_raises, assert_equal
+from numpy.testing import assert_allclose, assert_equal
 import numpy as np
-
 
 __author__ = "Doga Gursoy"
 __copyright__ = "Copyright (c) 2016, UChicago Argonne, LLC."
 __docformat__ = 'restructuredtext en'
 
-
-# Transformations
-
-def rotate_around_self():
-    P0 = Point([0, 0])
-    origin = P0
-    P1 = rotate(P0, np.pi/2, origin)
-    assert_allclose([P1.x, P1.y], [0, 0])
-
-
-def rotate_around_other():
-    P0 = Point([0, 0])
-    origin = Point([1, 0])
-    P1 = rotate(P0, np.pi/2, origin)
-    assert_allclose([P1.x, P1.y], [1, -1])
-
-
-def test_Point_rotate_around_self():
-    P0 = Point([0, 0])
-    origin = P0
-    P0.rotate(np.pi/2, origin)
-    assert_allclose([P0.x, P0.y], [0, 0])
-
-
-def test_Point_rotate_around_other():
-    P0 = Point([0, 0])
-    origin = Point([1, 0])
-    P0.rotate(np.pi/2, origin)
-    assert_allclose([P0.x, P0.y], [1, -1])
-
-    P0 = Point([0, 0])
-    origin = Point([0, 1])
-    P0.rotate(np.pi/2, origin)
-    assert_allclose([P0.x, P0.y], [1, 1])
-
-    P0 = Point([0, 0])
-    origin = Point([-1, 0])
-    P0.rotate(np.pi/2, origin)
-    assert_allclose([P0.x, P0.y], [-1, 1])
-
-    P0 = Point([0, 0])
-    origin = Point([-1, -1])
-    P0.rotate(np.pi/2, origin)
-    assert_allclose([P0.x, P0.y], [-2, 0])
-
-
-def test_translate():
-    P0 = Point([0, 0])
-    P0.translate([2.3, 4.5])
-    assert_equal([P0.x, P0.y], [2.3, 4.5])
-
-
-# Nonintersecting beams
-
-def test_beamcirc_nonintersecting_top():
-    circ = Circle(Point([0, 3]), 1)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 2)
-    assert_allclose(beamcirc(beam, circ), 0., rtol=1e-6)
-
-
-def test_beamcirc_nonintersecting_bottom():
-    circ = Circle(Point([0, -3]), 1)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 2)
-    assert_allclose(beamcirc(beam, circ), 0., rtol=1e-6)
-
-
-def test_beampoly_nonintersecting_top():
-    tri = Triangle(Point([0, 1]), Point([1, -1]), Point([-1, -1]))
-    beam = Probe(Point([-2, 2]), Point([2, 2]), 2)
-    assert_allclose(beampoly(beam, tri), 0., rtol=1e-6)
-
-
-def test_beampoly_nonintersecting_bottom():
-    tri = Triangle(Point([0, 1]), Point([1, -1]), Point([-1, -1]))
-    beam = Probe(Point([-2, -2]), Point([2, -2]), 2)
-    assert_allclose(beampoly(beam, tri), 0., rtol=1e-6)
-
-
-# Partial intersections
-
-def test_beamcirc_intersecting_partially_from_top_outside_center():
-    circ = Circle(Point([0, 1.5]), 1)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 2)
-    assert_allclose(beamcirc(beam, circ), 0.614184849304, rtol=1e-6)
-
-
-def test_beamcirc_intersecting_partially_from_bottom_outside_center():
-    circ = Circle(Point([0, -1.5]), 1)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 2)
-    assert_allclose(beamcirc(beam, circ), 0.614184849304, rtol=1e-6)
-
-
-def test_beamcirc_intersecting_partially_from_top_inside_center():
-    circ = Circle(Point([0, 0.5]), 1)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 2)
-    assert_allclose(beamcirc(beam, circ), 2.52740780429, rtol=1e-6)
-
-
-def test_beamcirc_intersecting_partially_from_bottom_inside_center():
-    circ = Circle(Point([0, -0.5]), 1)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 2)
-    assert_allclose(beamcirc(beam, circ), 2.52740780429, rtol=1e-6)
-
-
-def test_beampoly_intersecting_partially_from_top():
-    tri = Square(Point([0.5, 0.5]), side_length=1.0)
-    beam = Probe(Point([-2, 1]), Point([2, 1]), 1)
-    assert_allclose(beampoly(beam, tri), 1/2, rtol=1e-6)
-
-
-def test_beampoly_intersecting_partially_from_bottom():
-    tri = Square(Point([0.5, 0.5]), side_length=1.0)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 1)
-    assert_allclose(beampoly(beam, tri), 1/2, rtol=1e-6)
-
-
-# Full intersections
-
-def test_beamcirc_intersecting_fully_from_top_outside_center():
-    circ = Circle(Point([0, 1.5]), 3)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 2)
-    assert_allclose(beamcirc(beam, circ), 10.0257253792, rtol=1e-6)
-
-
-def test_beamcirc_intersecting_fully_from_bottom_outside_center():
-    circ = Circle(Point([0, -1.5]), 3)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 2)
-    assert_allclose(beamcirc(beam, circ), 10.0257253792, rtol=1e-6)
-
-
-def test_beamcirc_intersecting_fully_from_top_inside_center():
-    circ = Circle(Point([0, 0.5]), 3)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 2)
-    assert_allclose(beamcirc(beam, circ), 11.5955559562, rtol=1e-6)
-
-
-def test_beamcirc_intersecting_fully_from_bottom_inside_center():
-    circ = Circle(Point([0, -0.5]), 3)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 2)
-    assert_allclose(beamcirc(beam, circ), 11.5955559562, rtol=1e-6)
-
-
-def test_beamcirc_intersecting_fully():
-    circ = Circle(Point([0, 0]), 1)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 2)
-    assert_allclose(beamcirc(beam, circ), 3.14159265359, rtol=1e-6)
-
-
-def test_beampoly_intersecting_fully():
-    tri = Square(Point([0, 0]), side_length=2.0)
-    beam = Probe(Point([-2, 0]), Point([2, 0]), 3)
-    assert_allclose(beampoly(beam, tri), 4, rtol=1e-6)
-
-
-# Vertical intersection.
-
-def test_beamcirc_vertical_intersection():
-    circ = Circle(Point([0, 0]), 1)
-    beam = Probe(Point([-1, -1]), Point([1, 1]), 1)
-    assert_allclose(beamcirc(beam, circ), 1.91322295498, rtol=1e-6)
-
-
-def test_beampoly_vertical_intersection():
-    tri = Rectangle(Point([0, 0.5]), side_lengths=[10, 1])
-    beam = Probe(Point([0, -1]), Point([0, 1]), 1)
-    assert_allclose(beampoly(beam, tri), 1, rtol=1e-6)
-
-
-# Line
-
-def test_Line_slope_vertical():
-    line = Line(Point([0, -1]), Point([0, 1]))
-    assert_allclose(line.slope, np.inf, rtol=1e-6)
-
-
-def test_Line_yintercept_vertical():
-    line = Line(Point([0, -1]), Point([0, 1]))
-    assert_allclose(line.yintercept, np.inf, rtol=1e-6)
-
-
-def test_Line_slope():
-    line = Line(Point([-1, 0]), Point([1, 2]))
-    assert_allclose(line.slope, 1, rtol=1e-6)
-
-
-def test_Line_yintercept():
-    line = Line(Point([-1, 0]), Point([1, 2]))
-    assert_allclose(line.yintercept, 1, rtol=1e-6)
-
-
-def test_Line_same_points():
-    assert_raises(ValueError, Line, Point([1, 2]), Point([1, 2]))
-
-
-# Circle
 
 def test_Circle_area():
     circle = Circle(Point([0, 0]), 1)
