+++ tests/geometry/test_point.py
@@ -1,0 +1,115 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+# #########################################################################
+# Copyright (c) 2016, UChicago Argonne, LLC. All rights reserved.         #
+#                                                                         #
+# Copyright 2015. UChicago Argonne, LLC. This software was produced       #
+# under U.S. Government contract DE-AC02-06CH11357 for Argonne National   #
+# Laboratory (ANL), which is operated by UChicago Argonne, LLC for the    #
+# U.S. Department of Energy. The U.S. Government has rights to use,       #
+# reproduce, and distribute this software.  NEITHER THE GOVERNMENT NOR    #
+# UChicago Argonne, LLC MAKES ANY WARRANTY, EXPRESS OR IMPLIED, OR        #
+# ASSUMES ANY LIABILITY FOR THE USE OF THIS SOFTWARE.  If software is     #
+# modified to produce derivative works, such modified software should     #
+# be clearly marked, so as not to confuse it with the version available   #
+# from ANL.                                                               #
+#                                                                         #
+# Additionally, redistribution and use in source and binary forms, with   #
+# or without modification, are permitted provided that the following      #
+# conditions are met:                                                     #
+#                                                                         #
+#     * Redistributions of source code must retain the above copyright    #
+#       notice, this list of conditions and the following disclaimer.     #
+#                                                                         #
+#     * Redistributions in binary form must reproduce the above copyright #
+#       notice, this list of conditions and the following disclaimer in   #
+#       the documentation and/or other materials provided with the        #
+#       distribution.                                                     #
+#                                                                         #
+#     * Neither the name of UChicago Argonne, LLC, Argonne National       #
+#       Laboratory, ANL, the U.S. Government, nor the names of its        #
+#       contributors may be used to endorse or promote products derived   #
+#       from this software without specific prior written permission.     #
+#                                                                         #
+# THIS SOFTWARE IS PROVIDED BY UChicago Argonne, LLC AND CONTRIBUTORS     #
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT       #
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS       #
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL UChicago     #
+# Argonne, LLC OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,        #
+# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,    #
+# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;        #
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER        #
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT      #
+# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN       #
+# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE         #
+# POSSIBILITY OF SUCH DAMAGE.                                             #
+# #########################################################################
+
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
+from xdesign.geometry.point import *
+from numpy.testing import assert_allclose, assert_equal
+import numpy as np
+
+
+__author__ = "Doga Gursoy"
+__copyright__ = "Copyright (c) 2016, UChicago Argonne, LLC."
+__docformat__ = 'restructuredtext en'
+
+
+# Transformations
+
+def rotate_around_self():
+    P0 = Point([0, 0])
+    origin = P0
+    P1 = rotate(P0, np.pi/2, origin)
+    assert_allclose([P1.x, P1.y], [0, 0])
+
+
+def rotate_around_other():
+    P0 = Point([0, 0])
+    origin = Point([1, 0])
+    P1 = rotate(P0, np.pi/2, origin)
+    assert_allclose([P1.x, P1.y], [1, -1])
+
+
+def test_Point_rotate_around_self():
+    P0 = Point([0, 0])
+    origin = P0
+    P0.rotate(np.pi/2, origin)
+    assert_allclose([P0.x, P0.y], [0, 0])
+
+
+def test_Point_rotate_around_other():
+    P0 = Point([0, 0])
+    origin = Point([1, 0])
+    P0.rotate(np.pi/2, origin)
+    assert_allclose([P0.x, P0.y], [1, -1])
+
+    P0 = Point([0, 0])
+    origin = Point([0, 1])
+    P0.rotate(np.pi/2, origin)
+    assert_allclose([P0.x, P0.y], [1, 1])
+
+    P0 = Point([0, 0])
+    origin = Point([-1, 0])
+    P0.rotate(np.pi/2, origin)
+    assert_allclose([P0.x, P0.y], [-1, 1])
+
+    P0 = Point([0, 0])
+    origin = Point([-1, -1])
+    P0.rotate(np.pi/2, origin)
+    assert_allclose([P0.x, P0.y], [-2, 0])
+
+
+def test_translate():
+    P0 = Point([0, 0])
+    P0.translate([2.3, 4.5])
+    assert_equal([P0.x, P0.y], [2.3, 4.5])
+
+
+if __name__ == '__main__':
+    import nose
+    nose.runmodule(exit=False)
